home page work!
Improved overall visuals of the homepage
Added a last message view, hover over the short message to view the whole message.
Added a view to control the menu displayed to the user, users that are member will have the member menu available to them, this includes workers that are signed-up as members.
:)